[优化]: 使用clippy优化了代码结构与书写规范

+++ src-tauri/src/window_effect.rs
@@ -73,7 +73,7 @@ pub fn update_rounded_corners() -> Result<()> {
         };
         DwmSetWindowAttribute(
             hwnd,
-            DWMWINDOWATTRIBUTE(DWMWA_WINDOW_CORNER_PREFERENCE.0 as i32),
+            DWMWINDOWATTRIBUTE(DWMWA_WINDOW_CORNER_PREFERENCE.0),
             &corner_preference as *const _ as *const _,
             std::mem::size_of::<u32>() as u32,
         )
